+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-# Изменения
+# Лог решений
 
-  - 30.07.2022: [Who Like This](wholikethis/README.md)
+  - 30.07.2022: [Sum Of Numbers](sum_of_numbers/README.md)
   - 30.07.2022: [Find The Parity Outlier](parity_outlier/README.md)
+  - 30.07.2022: [Who Like This](wholikethis/README.md)